requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-CORS==3.0.10
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0

